requirements.txt
langchain
langchain-community
langchain-groq
langchain-chroma
langchain-huggingface
chromadb
pypdf
fastapi
uvicorn
streamlit
python-dotenv
sentence-transformers
pydantic
reportlab